Actualizacion ej 1 funciona boton editar
Ademas elimine algunos metodos de la clase gestionproducto y conexion que no se usaban

+++ TP-6/TP-6/Ejercicio1.aspx.cs
@@ -53,7 +53,7 @@ namespace TP_6
             Producto prod = new Producto();
             prod.NombreProducto = ((TextBox)gvProductos.Rows[e.RowIndex].FindControl("txt_NombreProducto")).Text;
             prod.CantidadPorUnidad = ((TextBox)gvProductos.Rows[e.RowIndex].FindControl("txt_CantidadPorUnidad")).Text;
-            prod.PrecioUnidad = Convert.ToDecimal(((TextBox)gvProductos.Rows[e.RowIndex].FindControl("txt_PrecioUnidad")).Text);
+            prod.PrecioUnidad = Convert.ToDouble(((TextBox)gvProductos.Rows[e.RowIndex].FindControl("txt_PrecioUnidad")).Text);
             prod.IdProducto = Convert.ToInt32(((Label)gvProductos.Rows[e.RowIndex].FindControl("lbl_idproducto")).Text);
             GestionProductos gProd = new GestionProductos();
